Switched out use of Int for Word8 in Binding

+++ src/Ermine/Pretty/Term.hs
@@ -106,13 +106,13 @@ prettyBody nm (Body ps gs wh) dvs vs kt kv =
  h pd gd (Just wd) = align $ nm <> pd <> nest 2 (softline <> gd)
                   <> nest 1 (line <> text "where" </> align wd)
 
- kv' (B (BodyDecl i)) _  = pure $ dvs !! i
+ kv' (B (BodyDecl i)) _  = pure $ dvs !! fromIntegral i
  kv' (B (BodyPat p)) _  = fromMaybe (error "PANIC: prettyBody: bad pattern variable reference")
                   $ pure . text <$> HM.lookup p bnd
- kv' (B (BodyWhere i)) _  = pure $ wvs !! i
+ kv' (B (BodyWhere i)) _  = pure $ wvs !! fromIntegral i
  kv' (F tm)    pr = prettyTerm tm rest pr kt kv
 
- kw (B (WhereDecl d)) = const $ pure $ dvs !! d
+ kw (B (WhereDecl d)) = const $ pure $ dvs !! fromIntegral d
  kw (B (WherePat p)) = fromMaybe (error "PANIC: prettyBody: bad pattern variable reference")
           $ const . pure . text <$> HM.lookup p bnd
  kw (F v) = kv v
